CONTRIBUTING: Fix hard-coded XnatUrl settings
When we change the url setting in 'Tools->Preferences->XNAT->XNAT URL', the browser will not visit the specified url.
Upon clicking the XNAT button on the top left toolbar, it'll still open a browser linking to the site https://central.xnat.org/

The hard-coded XnatUrl settings is fixed by reading the settings file.

Issue-Id: 224